+++ server/component/style/graphLegend/tpl_graphLegend.php
@@ -3,6 +3,6 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 ?>
-<div class="graphLegend">
+<div class="graphLegend <? echo $this->css; ?>">
     <?php $this->output_list_items(); ?>
 </div>